Count LZW bits in a long
`int totalBits = data.Length * 8` wraps negative at 256 MB, so the bit-scan
loop never ran and Decode returned an empty stream with no error. The input is
the previous filter stage's output, capped at MaxDecodedStreamBytes (512 MB
default) - twice the overflow threshold, so a 256-512 MB stage output was
accepted by policy and silently mis-decoded.

bitPos was an int too, overflowing at the identical boundary. Widening only
totalBits would have moved the silent failure rather than removed it: the loop
guard would pass, then bitPos would wrap mid-scan and index out of range.

The maxOutput guard at :164 was checked against the same question and is
sound - it runs every iteration, output.Length is genuinely what grows, and
per-iteration overshoot is bounded by the longest table entry.

AsciiHexFilter shares the shape at `new byte[data.Length * 2 + 1]` (overflows
at 1 GB), but it throws loudly rather than returning silence.

+++ Radzen.Blazor/Documents/Pdf/Objects/Filters/LzwFilter.cs
@@ -88,16 +88,18 @@ internal static class LzwFilter
         int width = 9;
         int nextCode = 258;
         int previous = -1;
-        int bitPos = 0;
+        long bitPos = 0;
 
-        int totalBits = data.Length * 8;
+        // 8 * int.MaxValue does not fit an int: a >= 256 MB stage output overflowed to a
+        // negative total, skipped the loop and returned an empty stream with no error.
+        long totalBits = (long)data.Length * 8;
 
         while (bitPos + width <= totalBits)
         {
             int code = 0;
             for (int i = 0; i < width; i++)
             {
-                int bit = (data[bitPos >> 3] >> (7 - (bitPos & 7))) & 1;
+                int bit = (data[bitPos >> 3] >> (7 - (int)(bitPos & 7))) & 1;
                 code = (code << 1) | bit;
                 bitPos++;
             }
